% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_repo.R
\name{build_repo}
\alias{build_repo}
\title{Build repo template}
\usage{
build_repo(github_page_url, site_author)
}
\arguments{
\item{github_page_url}{Your GitHub page name. You can check this when you
configure the GitHub pages in your repository. This is going to be written
in the mkdocs.yml file.}

\item{site_author}{The name of the author. This is going to be written in
the mkdocs.yml file.}
}
\description{
A function to create the minimum files and folders require to
build a repository to work with mkdocs and GitHub pages.
}
\details{
This will create an empty \code{docs} folder where all your md files
will be located coupled with output files as images. Also it will create a
\code{mkdocs.yml} file that uses material theme, contains the site url and the
author name.
}
\author{
Ronny Alexander Hernández Mora
}
