% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_files.R
\name{clean_files}
\alias{clean_files}
\title{Clean files in the docs folder that no longer exists in the root path.}
\usage{
clean_files(path)
}
\arguments{
\item{path}{The path where your directory with your notes are located. If
empty, it will take the path of the repository you are located.}
}
\description{
If you have erased no longer needed files in the root directory,
the render \code{md} files will continue to exists in the docs folder and hence,
they will continue to appear in your webpage. This function checks if there
are erased files in the root and delete them from the docs folder.
}
\examples{
\dontrun{
clean_files("~/Desktop/test_docmaker")
}

}
\author{
Ronny Alexander Hernández Mora
}
