% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_doc.R
\name{make_doc}
\alias{make_doc}
\title{Take your Rmarkdown file to mkdocs}
\usage{
make_doc(file, overwrite = TRUE, mkdocs_build = FALSE, mkdocs_deploy = FALSE)
}
\arguments{
\item{file}{The Rmarkdown file that we want to convert to markdown}

\item{mkdocs_build}{TRUE if you want to build the mkdocs documentation. This
will run the \verb{mkdocs build --config-file=mkdocs.yml} command}

\item{mkdocs_deploy}{TRUE if you want to deploy the mkdocs documentation. This
will run the \verb{mkdocs gh-deploy --strict --force} command}
}
\description{
This function will take your Rmarkdown file and transform it
to a markdown file that can be taken by mkdocs to render a HTML
}
\author{
Ronny Alexander Hernández Mora
}
