% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_doc.R
\name{publish_notes}
\alias{publish_notes}
\title{Convert from Rmarkdown to markdown}
\usage{
publish_notes(
  file,
  overwrite = TRUE,
  mkdocs_build = FALSE,
  mkdocs_deploy = FALSE
)
}
\arguments{
\item{file}{The Rmarkdown file that we want to convert to markdown}

\item{mkdocs_build}{TRUE if you want to build the mkdocs documentation. This
will run the \verb{mkdocs build --config-file=mkdocs.yml} command}
}
\description{
This function will take the Rmarkdown file and transform it
to a markdown file that can be taken by mkdocs to render a HTML
}
\author{
Ronny Alexander Hernández Mora
}
