% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_all_docs.R
\name{render_docs}
\alias{render_docs}
\title{Render and deploy all your Rmd documents}
\usage{
render_docs(deploy = FALSE)
}
\arguments{
\item{deploy}{If TRUE all the Rmd documents will be deploy to your site}
}
\description{
Take all your Rmarkdown documents and render to markdown. This
function also can deploy all your content at once to your GitHub site.
}
\author{
Ronny Alexander Hernández Mora
}
